common oled code for k3
