+++ keyboards/ergohaven/k03/rules.mk
@@ -28,3 +28,5 @@ AUTO_SHIFT_ENABLE = yes
 SRC += keyboards/ergohaven/ergohaven_rgb.c
 SRC += keyboards/ergohaven/ergohaven.c
 SRC += keyboards/ergohaven/lang_ru_en.c
+SRC += keyboards/ergohaven/ergohaven_oled.c
+SRC += oled/ergohaven_dark.c
